requirements.txt
streamlit
st-gsheets-connection
datetime
pandas
pickle
plotly.express
StandardScaler
time